+++ src/common/libsubprocess/Makefile.am
@@ -80,6 +80,7 @@ test_ldadd = \
         $(top_builddir)/src/common/libtap/libtap.la \
         $(top_builddir)/src/common/libsubprocess/libsubprocess.la \
         $(top_builddir)/src/common/libflux-core.la \
+        $(top_builddir)/src/common/libflux/libflux.la \
         $(top_builddir)/src/common/libflux-internal.la
 
 test_ldflags = \
