+++ t/Makefile.am
@@ -749,14 +749,16 @@ reactor_reactorcat_CPPFLAGS = $(test_cppflags)
 reactor_reactorcat_LDADD = $(test_ldadd)
 reactor_reactorcat_LDFLAGS = \
 	$(top_builddir)/src/common/libsubprocess/libsubprocess.la \
-	$(test_ldflags)
+	$(test_ldflags) \
+	$(top_builddir)/src/common/libflux/libflux.la
 
 rexec_rexec_SOURCES = rexec/rexec.c
 rexec_rexec_CPPFLAGS = $(test_cppflags)
 rexec_rexec_LDADD = $(test_ldadd)
 rexec_rexec_LDFLAGS = \
 	$(top_builddir)/src/common/libsubprocess/libsubprocess.la \
-	$(test_ldflags)
+	$(test_ldflags) \
+	$(top_builddir)/src/common/libflux/libflux.la
 
 rexec_rexec_count_stdout_SOURCES = rexec/rexec_count_stdout.c
 rexec_rexec_count_stdout_CPPFLAGS = $(test_cppflags)
@@ -768,7 +770,8 @@ rexec_rexec_getline_CPPFLAGS = $(test_cppflags)
 rexec_rexec_getline_LDADD = $(test_ldadd)
 rexec_rexec_getline_LDFLAGS = \
 	$(top_builddir)/src/common/libsubprocess/libsubprocess.la \
-	$(test_ldflags)
+	$(test_ldflags) \
+	$(top_builddir)/src/common/libflux/libflux.la
 
 ingest_job_manager_la_SOURCES = ingest/job-manager.c
 ingest_job_manager_la_CPPFLAGS = $(test_cppflags)
